RR-Team_66-yadtq-yet-another-distributed-task-queue-/requirements.txt
attrs==21.2.0
bcrypt==3.2.0
blinker==1.4
cachetools==5.0.0
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
cryptography==3.4.8
dbus-python==1.2.18
defer==1.0.4
distro==1.7.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==3.0.2
fasteners==0.14.1
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.1.0
greenlet==1.1.2
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
jsonschema==3.2.0
kafka-python==2.0.2
keyring==23.5.0
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
ptyprocess==0.7.0
py4j==0.10.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycups==2.0.1
PyDrive==1.3.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
pyspark==3.5.3
python-dotenv==0.19.2
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
rsa==4.8
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==1.4.31
texttable==1.6.4
uritemplate==3.0.1
urllib3==1.26.5
websocket-client==1.2.3
xdg==5
zipp==1.0.0

